% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathTS.R
\name{ApproxPath.TS.Para}
\alias{ApproxPath.TS.Para}
\title{This function helps to do ApproxPath.TS in parallel}
\usage{
ApproxPath.TS.Para(mat, ...)
}
\arguments{
\item{mat:}{A matrix which contains X and Y}

\item{...:}{args passed to ApproxPath.TS}
}
\value{
a vector of test statistic for each test specified
}
\description{
This function could be called in parallel
}
