% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproxTS.R
\name{ApproxPath.TS}
\alias{ApproxPath.TS}
\title{Approximate Calculation of Leave-One-Covariate-Out regularization path test staitsic}
\usage{
ApproxPath.TS(X, Y, which.covariate, betaNull, multiTest,
  path.method = "lars", norm = "L2.squared", normalize = TRUE,
  intercept = FALSE, ridgePen = 0, numInterpolation = 10000)
}
\arguments{
\item{X:}{a n by p matrix, design matrix}

\item{Y:}{a n by 1 matrx, response}

\item{which.covariate:}{a vector or a list of vector, specify which covaritate \eqn{ \beta_j } to perform our test}

\item{betaNull:}{a vector or a list of vector, should be the same shape as which.covariate. 
Specify the Null hypothesis, \eqn{H_0: \beta_j = betaNull vs H_1 \neq betaNull }
    if using multiTest = FALSE, then will return}

\item{multiTest:}{boolean, TRUE: test simultaneously, FALSE: test individually}

\item{path.method:}{string, c("lars", "plus.lasso", "plus.mc+", "plus.scad" )}

\item{norm:}{string, c("L1", "L2", "L2.squared", "L_inf")}

\item{normalize:}{boolean, using normalization or not}

\item{intercept:}{boolean, include intercept or not}
}
\value{
a vector of test statistic for each test specified
}
\description{
This function performs the approximate Leave-One-Covariate-Out regularization path estimator using interpolation and approximate integration
}
\examples{
X = matrix(rnorm(100*12), 100, 12)
beta = c(1,2, rep(0, 10))
Y = X \%*\% beta + rnorm(100)
# Test \\eqn{H_0: \\beta_1 = 1} and \\eqn{H_0: \\beta_1 = 2} individually 
ApproxPath.TS(X, Y, which.covariate = c(1, 2), betaNull = c(1, 2), multiTest = FALSE)
# Test \\eqn{H_0: \\beta_1 = 1, \\beta_1 = 2} simultaneously
ApproxPath.TS(X, Y, which.covariate = list(c(1, 2)), betaNull = list(c(1, 2)), multiTest = TRUE)
# Test \\eqn{H_0: \\beta_1 = 1, \\beta_2 = 2} simultaneously and test \\eqn{H_0: \\beta_1 = 0, \\beta_2 = 0}
ApproxPath.TS(X, Y, which.covariate = list(c(1, 2), c(1, 2)), betaNull = list(c(1, 2), c(0, 0)), multiTest = TRUE)
}
