% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{dataGen}
\alias{dataGen}
\title{Generate data with design specified}
\usage{
dataGen(setting = "dep", ...)
}
\arguments{
\item{setting:}{c("dep")}

\item{...:}{args passed to depenDesign}
}
\value{
a list with X and Y
}
\description{
This function wraps up different design for the simulated data
}
