% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{depenDesign}
\alias{depenDesign}
\title{Generate data with correlated design}
\usage{
depenDesign(n, p, beta, rho)
}
\arguments{
\item{n:}{integar, sample size}

\item{p:}{integar, number of covariates}

\item{rho:}{correlation, if real value, $eqn{Sigma_{i,j} = rho ^ {|i-j|}}
if string, c("equl", "weak_equl")}
}
\value{
a list with X and Y
}
\description{
This function Generate the data for simulation
}
