% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathTS_Power.R
\name{Path.Resample.Power}
\alias{Path.Resample.Power}
\title{Simulate the power of LOCO regularization path test statistics}
\usage{
Path.Resample.Power(n = 100, p = 1000, beta = c(rep(1, 10), rep(0,
  990)), rho = 0.5, iter = 500, B = 500, setting = "dep",
  which.covariate = 1, betaNull = 1, multiTest = FALSE, ...)
}
\arguments{
\item{n:}{integar, sample size}

\item{p:}{integar, number of covariates}

\item{rho:}{correlation, if real value, $eqn{Sigma_{i,j} = rho ^ {|i-j|}}
if string, c("equl", "weak_equl")}

\item{beta:}{true beta coefficient}

\item{iter:}{number of iterations}

\item{B:}{number of bootstraps}

\item{which.covariate:}{args passed to Path.TS}

\item{betaNull:}{args passed to Path.TS}

\item{multiTest:}{args passed to Path.TS}

\item{...:}{args passed to Path.TS}
}
\value{
a vector of power for each Hypothesis test specified
}
\description{
This function simulate the power of LOCO regularization path test statistics
}
