% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleTS_Engine.R
\name{Path.Resample.Process}
\alias{Path.Resample.Process}
\title{The engine to perform residual bootsrrap of the Leave-One-Covariate-Out regularization path test staitsic}
\usage{
Path.Resample.Process(X, Y, multiTest, residual, b.Null, beta.index,
  betaNull, B = 500, exact, parallel = FALSE, ...)
}
\arguments{
\item{X:}{a n by p matrix, design matrix}

\item{Y:}{a n by 1 matrx, response}

\item{multiTest:}{boolean, TRUE: test simultaneously, FALSE: test individually}

\item{betaNull:}{same args for ExactPath.TS}

\item{B:}{integer, number of bootstraps}

\item{parallel:}{boolean, run the whole bootstrap process in parallell, useful if run in a multicore cpu, especially in cluster}

\item{exact:}{boolean, using ExactPath.TS or AprroxPath.TS}

\item{b.Null:}{Args from Path.Resample}

\item{beta.index:}{Args from Path.Resample}

\item{...:}{args wull be passed to ExactPath.TS or ApproxPath.Ts}
}
\value{
Will be passed to Path.Resample
}
\description{
This function performs residual bootsrrap the Leave-One-Covariate-Out regularization path test staitsic
}
