% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleTS.R
\name{Path.Resample}
\alias{Path.Resample}
\title{Residual bootsrrap the Leave-One-Covariate-Out regularization path test staitsic}
\usage{
Path.Resample(X, Y, which.covariate, betaNull, multiTest, B = 500,
  parallel = FALSE, exact = TRUE, beta.init = "adaptive",
  beta.true = beta, ...)
}
\arguments{
\item{X:}{a n by p matrix, design matrix}

\item{Y:}{a n by 1 matrx, response}

\item{which.covariate:}{a vector or a list of vector, specify which covaritate \eqn{ \beta_j } to perform our test}

\item{betaNull:}{a vector or a list of vector, should be the same shape as which.covariate. 
Specify the Null hypothesis, \eqn{H_0: \beta_j = betaNull vs H_1 \neq betaNull }
    if using multiTest = FALSE, then will return}

\item{multiTest:}{boolean, TRUE: test simultaneously, FALSE: test individually}

\item{B:}{integer, number of bootstraps}

\item{parallel:}{boolean, run the whole bootstrap process in parallell, useful if run in a multicore cpu, especially in cluster}

\item{exact:}{boolean, using ExactPath.TS or AprroxPath.TS}

\item{beta.init:}{str, c("adaptive", "de-sparse", "MC+", "SCAD", "Truth"). Under developing, please use "adaptive" and "Truth". 
Needed if perform high-dimensional regression, low-dimensional regression will use LSE.}

\item{beta.true:}{true value of beta, if your beta.init = "Truth", need to specify this}

\item{...:}{args wull be passed to ExactPath.TS or ApproxPath.Ts}
}
\value{
a list of p-value, TS, TS_null, rej or nor
}
\description{
This function performs residual bootsrrap the Leave-One-Covariate-Out regularization path test staitsic and would return test p-value
}
\examples{
X = matrix(rnorm(100*12), 100, 12)
beta = c(1,2, rep(0, 10))
Y = X \%*\% beta + rnorm(100)
# Test \\eqn{H_0: \\beta_1 = 1} and \\eqn{H_0: \\beta_1 = 2} individually 
Path.Resample(X, Y, which.covariate = c(1, 2), betaNull = c(1, 2), multiTest = FALSE, B = 100)
# Test \\eqn{H_0: \\beta_1 = 1, \\beta_1 = 2} simultaneously
Path.Resample(X, Y, which.covariate = list(c(1, 2)), betaNull = list(c(1, 2)), multiTest = TRUE, B = 100)
# Test \\eqn{H_0: \\beta_1 = 1, \\beta_2 = 2} simultaneously and test \\eqn{H_0: \\beta_1 = 0, \\beta_2 = 0}
Path.Resample(X, Y, which.covariate = list(c(1, 2), c(1, 2)), betaNull = list(c(1, 2), c(0, 0)), multiTest = TRUE, B = 100)
}
