% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathTS.R
\name{Path.TS.Para}
\alias{Path.TS.Para}
\title{This function wraps up ApproxPath.TS.Para and ExactPath.Ts.Para}
\usage{
Path.TS.Para(exact = TRUE, ...)
}
\arguments{
\item{exact:}{boolean, using exact calculation or not}

\item{...:}{args passed to ExactPath.TS and ApproxPath.TS}
}
\value{
a vector of test statistic for each test specified
}
\description{
This function could be called in parallel
}
