% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathTS.R
\name{Path.TS}
\alias{Path.TS}
\title{This function wraps up ApproxPath.TS and ExactPath.Ts}
\usage{
Path.TS(exact = TRUE, ...)
}
\arguments{
\item{exact:}{boolean, using exact calculation or not}

\item{...:}{args passed to ExactPath.TS and ApproxPath.TS}
}
\value{
a vector of test statistic for each test specified
}
\description{
This function wraps up ApproxPath.TS and ExactPath.Ts
}
